Reorganize CTC zh-CN scripts into logical directories
Create organized directory structure:
- conversion/ - Export and quantization scripts
- validation/ - Validation scripts (compare vs NeMo reference)
- benchmark/ - Performance benchmarking scripts

Moved files:
- conversion/export-ctc-zh-cn.py (decoder-only export)
- conversion/export-full-pipeline.py (full pipeline export)
- conversion/quantize-encoder-advanced.py (post-training quantization)
- conversion/individual_components.py (PyTorch wrappers)
- validation/validate-ctc-zh-cn.py (single-sample validation)
- validation/benchmark-cer.py (multi-sample validation)
- benchmark/benchmark-full-pipeline.py (pure CoreML benchmark)
- benchmark/text_normalizer.py (shared utility)

Updated documentation (README.md, CONVERSION_NOTES.md) with new paths
and directory structure.
