+++ models/stt/parakeet-ctc-0.6b-zh-cn/coreml/benchmark/benchmark-full-pipeline.py
