+++ models/stt/parakeet-ctc-0.6b-zh-cn/coreml/conversion/individual_components.py
