+++ models/stt/parakeet-ctc-0.6b-zh-cn/coreml/validation/validate-ctc-zh-cn.py
